# packages/api-contracts/rest/openapi.yaml
openapi: 3.1.0
info:
  title: ManaBandhu REST API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /chat/conversations:
    get:
      operationId: listConversations
      responses:
        '200':
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
    post:
      operationId: createConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationInput'
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
  /chat/conversations/{id}:
    get:
      operationId: getConversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Conversation detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
  /chat/conversations/{id}/messages:
    get:
      operationId: listMessages
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      operationId: sendMessage
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageInput'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /chat/conversations/{id}/participants:
    get:
      operationId: listParticipants
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConversationParticipant'
    post:
      operationId: addParticipant
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddParticipantInput'
      responses:
        '201':
          description: Participant added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationParticipant'
  /assistant/conversations:
    get:
      operationId: listAssistantConversations
      responses:
        '200':
          description: Assistant conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssistantConversation'
    post:
      operationId: createAssistantConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantConversation'
  /assistant/conversations/{id}/messages:
    get:
      operationId: listAssistantMessages
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Assistant messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssistantMessage'
    post:
      operationId: sendAssistantMessage
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendAssistantMessageInput'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantMessage'
  /assistant/conversations/{id}/stream:
    get:
      operationId: streamAssistantMessages
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Streaming assistant response
          content:
            text/event-stream:
              schema:
                type: string
  /assistant/messages/{id}/citations:
    get:
      operationId: listAssistantCitations
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Citations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssistantCitation'
    post:
      operationId: createAssistantCitation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssistantCitationInput'
      responses:
        '201':
          description: Citation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantCitation'
...
  /events:
    get:
      operationId: listEvents
      parameters:
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Event' }
    post:
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEventInput' }
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
...
  /events/{eventId}:
    get:
      operationId: getEvent
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Event detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
  /events/{eventId}/attendance:
    get:
      operationId: listEventAttendance
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Event attendance
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EventAttendance' }
    post:
      operationId: createEventAttendance
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEventAttendanceInput' }
      responses:
        '201':
          description: Attendance created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventAttendance' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Conversation:
      type: object
      additionalProperties: false
      required: [id, type, createdAt]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [DIRECT, GROUP] }
        title: { type: string, maxLength: 200 }
        createdAt: { type: string, format: date-time }
        lastMessageAt: { type: string, format: date-time }
    CreateConversationInput:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type: { type: string, enum: [DIRECT, GROUP] }
        title: { type: string, maxLength: 200 }
    Message:
      type: object
      additionalProperties: false
      required: [id, conversationId, senderId, body, messageType, createdAt]
      properties:
        id: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        senderId: { type: string, format: uuid }
        body: { type: string, maxLength: 4000 }
        messageType: { type: string, enum: [TEXT, IMAGE, SYSTEM] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SendMessageInput:
      type: object
      additionalProperties: false
      required: [body, messageType]
      properties:
        body: { type: string, maxLength: 4000 }
        messageType: { type: string, enum: [TEXT, IMAGE, SYSTEM] }
    ConversationParticipant:
      type: object
      additionalProperties: false
      required: [id, conversationId, userId, role, joinedAt]
      properties:
        id: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        joinedAt: { type: string, format: date-time }
        leftAt: { type: string, format: date-time }
        role: { type: string, enum: [OWNER, MEMBER] }
    AddParticipantInput:
      type: object
      additionalProperties: false
      required: [userId]
      properties:
        userId: { type: string, format: uuid }
    AssistantConversation:
      type: object
      additionalProperties: false
      required: [id, userId, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        title: { type: string, maxLength: 200 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateConversationRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title: { type: string, maxLength: 200 }
    AssistantMessage:
      type: object
      additionalProperties: false
      required: [id, conversationId, role, content, createdAt]
      properties:
        id: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        role: { type: string, enum: [USER, ASSISTANT, SYSTEM] }
        content: { type: string, maxLength: 4000 }
        createdAt: { type: string, format: date-time }
    SendAssistantMessageInput:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        content: { type: string, maxLength: 4000 }
    AssistantCitation:
      type: object
      additionalProperties: false
      required: [id, messageId, title, snippet, createdAt]
      properties:
        id: { type: string, format: uuid }
        messageId: { type: string, format: uuid }
        title: { type: string, maxLength: 300 }
        url: { type: string, format: uri, maxLength: 500 }
        snippet: { type: string, maxLength: 4000 }
        createdAt: { type: string, format: date-time }
    CreateAssistantCitationInput:
      type: object
      additionalProperties: false
      required: [title, snippet]
      properties:
        title: { type: string, maxLength: 300 }
        url: { type: string, format: uri, maxLength: 500 }
        snippet: { type: string, maxLength: 4000 }
...
    Event:
      type: object
      additionalProperties: false
      required: [id, organizerId, title, location, startAt, endAt, categoryId, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        organizerId: { type: string, format: uuid }
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        location: { type: string, maxLength: 200 }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        categoryId: { type: string, format: uuid }
        status: { type: string, enum: [DRAFT, PUBLISHED, CANCELED, COMPLETED] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateEventInput:
      type: object
      additionalProperties: false
      required: [title, location, startAt, endAt, categoryId]
      properties:
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        location: { type: string, maxLength: 200 }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        categoryId: { type: string, format: uuid }
    CreateEventAttendanceInput:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string, enum: [INTERESTED, GOING, NOT_GOING] }
    EventAttendance:
      type: object
      additionalProperties: false
      required: [id, eventId, userId, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        eventId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        status: { type: string, enum: [INTERESTED, GOING, NOT_GOING] }
        createdAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []